skills: the great rewrite of the lvl2's
Based on the level 1 rewrites, I focused heavily on keeping the lvl2 skills to be a lot LIKE your lvl1's, but with slightly different deviations. I went with structure first, then I went to naming them. It's a huge change, but a great one I think.

This gets you used to a certain skill, maybe even training you to like certain numbers as you roll them. Also helps with collisions GREATLY.

Some of the OP moves have been nerfed. Jab is gone, and there are much fewer "Redirects" and such.

The basic idea is that the 3x lvl1 moves that get either more ridiculous or they balance out with "seasoning". So Hit 'n' Run either turns into Assault (with Rampage being a signature move), or Blast (with Detonate hitting all security tokens at once).

Interestingly, the names are much more apt now - even though before I started with the names and then went to structure, this time the names work much better. And they also evoke the idea of leveling up.

My favorite joke in this, though, is that "Unplug" is the most powerful Camera-related move. It's like you need tons of experience to learn how to do the obvious.